[Kernel][Helion] Consolidate generated kernel dispatch
