+++ vllm/kernels/helion_generated/__init__.py
@@ -1,3 +1,25 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 """Runtime support for checked-in kernels generated by Helion."""
+
+from __future__ import annotations
+
+from collections.abc import Iterable
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    import torch
+
+
+def warm_up_helion_kernels(
+    token_counts: Iterable[int],
+    device: torch.device | str = "cuda",
+) -> None:
+    from vllm.kernels.helion_generated.dispatcher import (
+        warmup_generated_fusion_kernels,
+        warmup_per_token_group_fp8_quant,
+    )
+
+    token_counts = tuple(token_counts)
+    warmup_per_token_group_fp8_quant(token_counts, device)
+    warmup_generated_fusion_kernels(token_counts, device)
